Extract xtras crate
These extensions to `xtra` are only used in `daemon` at the moment,
but they are definitely unrelated to the functionality offered by the
`daemon` crate.
