+++ daemon/src/connection.rs
@@ -1,5 +1,3 @@
-use crate::address_map::AddressMap;
-use crate::address_map::Stopping;
 use crate::collab_settlement_taker;
 use crate::future_ext::FutureExt;
 use crate::model::cfd::OrderId;
@@ -15,8 +13,6 @@ use crate::wire;
 use crate::wire::EncryptedJsonCodec;
 use crate::wire::TakerToMaker;
 use crate::wire::Version;
-use crate::xtra_ext::LogFailure;
-use crate::xtra_ext::SendInterval;
 use crate::Tasks;
 use anyhow::bail;
 use anyhow::Context;
@@ -35,6 +31,10 @@ use tokio_util::codec::Framed;
 use xtra::prelude::MessageChannel;
 use xtra::KeepRunning;
 use xtra_productivity::xtra_productivity;
+use xtras::address_map::Stopping;
+use xtras::AddressMap;
+use xtras::LogFailure;
+use xtras::SendInterval;
 
 /// Time between reconnection attempts
 const CONNECT_TO_MAKER_INTERVAL: Duration = Duration::from_secs(5);
