+++ xtras/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "xtras"
+version = "0.1.0"
+edition = "2021"
+description = "Extension trait for the xtra actor framework"
+
+[dependencies]
+anyhow = "1"
+async-trait = "0.1.52"
+regex = "1"
+thiserror = "1"
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "time"] }
+tracing = { version = "0.1" }
+xtra = { version = "0.6" }
