+++ xtras/src/lib.rs
@@ -1,0 +1,11 @@
+mod actor_name;
+pub mod address_map;
+mod log_failure;
+mod send_async_safe;
+mod send_interval;
+
+pub use actor_name::ActorName;
+pub use address_map::AddressMap;
+pub use log_failure::LogFailure;
+pub use send_async_safe::SendAsyncSafe;
+pub use send_interval::SendInterval;
